+++ simplecms/lib/constants.py
@@ -10,3 +10,4 @@ import os
 THISDIR = os.path.dirname(os.path.abspath(__file__))
 APPDIR = os.path.abspath(os.path.join(THISDIR, ".."))
 HTTPROOTDIR = os.path.join(APPDIR, "http_root")
+